feat(#25): partial solution for rabbitmq has been added
