+++ Docker/.env
@@ -2,3 +2,7 @@ COMPOSE_CONVERT_WINDOWS_PATHS=1
 MYSQL_USER=user
 MYSQL_PASSWORD=user
 MYSQL_DB=users
+RABBIT_HOST=project_rabbit
+RABBIT_PORT=5672
+RABBIT_USER=guest
+RABBIT_PASS=guest
