+++ Docker/docker-compose.dev.yml
@@ -105,6 +105,10 @@ services:
       - ADMIN_PASS=password
       - JWT_SECRET=6f93a7420057e6e71ced 
       - spring_profiles_active=prod
+      - RABBIT_HOST=${RABBIT_HOST}
+      - RABBIT_PORT=${RABBIT_PORT}
+      - RABBIT_USER=${RABBIT_USER}
+      - RABBIT_PASS=${RABBIT_PASS}
     networks:
       - project-network
     ports:
