+++ ProjectService/ToDoAPI/Controllers/ToDoController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
 using Newtonsoft.Json;
-using RabbitMQ.Client;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -39,7 +38,6 @@ namespace ToDoAPI.Controllers
             try
             {
                 await _repository.Create(toDo);
-                sendTaskCreatedMessage(toDo);
 
             }
             catch (MongoWriteException e)
@@ -139,29 +137,6 @@ namespace ToDoAPI.Controllers
             await _repository.Delete(id);
 
             return Ok();
-        }
-
-        private void sendTaskCreatedMessage(ToDo toDoToSend)
-        {
-            var factory = new ConnectionFactory() { HostName = "localhost" };
-            using (var connection = factory.CreateConnection())
-            using (var channel = connection.CreateModel())
-            {
-                channel.QueueDeclare(queue: "hello",
-                                     durable: false,
-                                     exclusive: false,
-                                     autoDelete: false,
-                                     arguments: null);
-
-                string message = JsonConvert.SerializeObject(toDoToSend);
-                var body = Encoding.UTF8.GetBytes(message);
-
-                channel.BasicPublish(exchange: "",
-                                     routingKey: "hello",
-                                     basicProperties: null,
-                                     body: body);
-                Console.WriteLine(" [x] Sent {0}", message);
-            }
         }
     }
 }
