+++ UserService/src/main/java/com/example/service/RabbitMqService.java
@@ -1,0 +1,22 @@
+package com.example.service;
+
+import com.example.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class RabbitMqService {
+
+    private final RabbitTemplate rabbitTemplate;
+    private final UserRepository userRepository;
+
+    @RabbitListener(queues = "project_queue")
+    public void getProjectString(String message) {
+        log.debug("From queue: {}", message);
+    }
+}
