+++ UserService/src/main/resources/application.yml
@@ -4,3 +4,8 @@ spring:
       matching-strategy: ant_path_matcher
   jpa:
     show-sql: true
+  rabbitmq:
+    host: localhost
+    port: 5672
+    username: guest
+    password: guest
